chore(shared): move useGroupFields to shared
